Error fixed display post, post delete button; add post timestamp by using moment.js
