+++ helpers/index.js
@@ -1,0 +1,9 @@
+const moment = require("moment");
+
+const formatDate = (date) => moment(date).format("M/D/YYYY");
+
+const helpers = {
+  formatDate,
+};
+
+module.exports = helpers;
