+++ public/js/post.js
@@ -1,4 +1,4 @@
-const createFormHandler = async (event) => {
+const createPostHandler = async (event) => {
   event.preventDefault();
 
   //   Eingabefelder habe ich so ausgelesen
@@ -15,29 +15,13 @@ const createFormHandler = async (event) => {
   });
   if (response.ok) {
     document.location.replace("/dashboard");
+  } else {
+    console.log("New post failed.", response);
   }
   //   evtl. Error Fall noch einbauen.
 };
 
-const onClickDelete = async (event) => {
-  event.preventDefault;
-  const id = event.currentTarget.id;
-  const options = {
-    method: "DELETE",
-    redirect: "follow",
-  };
-  const response = await fetch(`/api/post/${id}`, options);
-
-  if (response.status !== 200) {
-    console.info("Failed to delete post.");
-  } else {
-    window.location.replace("/dashboard");
-  }
-};
-
-$(`[name="delete-post"]`).click(onClickDelete);
-
 document
   .querySelector(".create-form")
   // Konfiguration des Buttons; der Button soll darauf hören, wenn ein submit kommt. Erst dann wird die Funktion "createFormHandler" ausgeführt.
-  .addEventListener("submit", createFormHandler);
+  .addEventListener("submit", createPostHandler);
