+++ server.js
@@ -4,8 +4,9 @@ const path = require("path");
 const routes = require("./controllers");
 const sequelize = require("./config/connection");
 const session = require("express-session");
+const helpers = require("./helpers");
 
-const hbs = exphbs.create({});
+const hbs = exphbs.create({ helpers });
 const app = express();
 const PORT = process.env.PORT || 3001;
