+++ views/layouts/main.handlebars
@@ -2,6 +2,12 @@
   <head>
     <meta charset="UTF-8" />
     <title>Tech-Blog</title>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.slim.min.js"
+    ></script>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.0/js/bootstrap.min.js"
+    ></script>
     {{! <link rel="stylesheet" href="/css/jass.css"> }}
     {{! <link rel="stylesheet" href="/css/style.css"> }}
     <link
@@ -22,11 +28,5 @@
       </main>
     </div>
   </body>
-  <script
-    src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.slim.min.js"
-  ></script>
-  <script
-    src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.5.0/js/bootstrap.min.js"
-  ></script>
-  <script src="/js/post.js"></script>
+
 </html>
